+++ examples/vite.config.ts
@@ -1,8 +1,9 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+import { viteSingleFile } from 'vite-plugin-singlefile';
 
 export default defineConfig({
   base: './',
-  plugins: [react()],
+  plugins: [react(), viteSingleFile()],
   resolve: { dedupe: ['react', 'react-dom'] },
 });